+++ tests/trestle/core/control_io_test.py
@@ -171,9 +171,6 @@ end of text
 
     context = ControlContext.generate(ContextPurpose.CATALOG, True, tmp_path, tmp_path)
     writer = ControlWriter()
-    # writer.write_control_for_editing(
-    #     tmp_path, control, 'My Group Title', None, None, additional_content, False, None, False, None, None
-    # )
     writer.write_control_for_editing(context, control, tmp_path, 'My Group Title')
 
     md_path = tmp_path / f'{control.id}.md'
@@ -201,9 +198,6 @@ def test_control_objective(tmp_path: pathlib.Path) -> None:
     # write it out as markdown in a separate directory to avoid name clash
     context = ControlContext.generate(ContextPurpose.CATALOG, True, tmp_path, sub_dir)
     control_writer = ControlWriter()
-    # control_writer.write_control_for_editing(
-    #     sub_dir, control, 'My Group Title', None, None, False, False, None, False, None, None
-    # )
     control_writer.write_control_for_editing(context, control, sub_dir, group_title)
     # confirm the newly written markdown text is identical to what was read originally
     assert test_utils.text_files_equal(md_path, sub_dir / 'xy-9.md')
@@ -353,24 +347,24 @@ def test_get_control_param_dict(tmp_trestle_dir: pathlib.Path) -> None:
     catalog = ProfileResolver.get_resolved_profile_catalog(tmp_trestle_dir, prof_a_path)
     catalog_interface = CatalogInterface(catalog)
     control = catalog_interface.get_control('ac-1')
-    param_dict = ControlReader.get_control_param_dict(control, False)
+    param_dict = ControlInterface.get_control_param_dict(control, False)
     # confirm profile value is used
-    assert ControlReader.param_values_as_str(param_dict['ac-1_prm_1']) == 'all alert personnel'
+    assert ControlInterface.param_values_as_str(param_dict['ac-1_prm_1']) == 'all alert personnel'
     # confirm original param label is used since no value was assigned
-    assert ControlReader.param_to_str(
+    assert ControlInterface.param_to_str(
         param_dict['ac-1_prm_7'], ParameterRep.VALUE_OR_LABEL_OR_CHOICES
     ) == 'organization-defined events'
     param = control.params[0]
     param.values = None
     param.select = common.ParameterSelection(how_many=common.HowMany.one_or_more, choice=['choice 1', 'choice 2'])
-    param_dict = ControlReader.get_control_param_dict(control, False)
-    assert ControlReader.param_to_str(
+    param_dict = ControlInterface.get_control_param_dict(control, False)
+    assert ControlInterface.param_to_str(
         param_dict['ac-1_prm_1'], ParameterRep.VALUE_OR_LABEL_OR_CHOICES
     ) == 'choice 1; choice 2'
-    assert ControlReader.param_to_str(
+    assert ControlInterface.param_to_str(
         param_dict['ac-1_prm_1'], ParameterRep.VALUE_OR_LABEL_OR_CHOICES, True
     ) == 'Choose one or more: choice 1; choice 2'
-    assert ControlReader.param_to_str(
+    assert ControlInterface.param_to_str(
         param_dict['ac-1_prm_1'], ParameterRep.VALUE_OR_LABEL_OR_CHOICES, True, True
     ) == 'Choose one or more: [choice 1; choice 2]'
 
@@ -407,10 +401,6 @@ def test_write_control_header_params(overwrite_header_values, tmp_path: pathlib.
     context.yaml_header = header
     context.overwrite_header_values = overwrite_header_values
     control_writer = ControlWriter()
-    # write the control back out with the test header
-    # control_writer.write_control_for_editing(
-    #     tmp_path, orig_control_read, group_title, header, None, False, False, None, overwrite_header_values, None, None
-    # )
     control_writer.write_control_for_editing(context, orig_control_read, tmp_path, group_title)
     # header_2 should have 2 params: 3 and 4
     header_2, _ = markdown_processor.read_markdown_wo_processing(control_path)
